LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/transformations/2d.transformation.rotate.direction-expected.txt

2d.transformation.rotate.direction

rotate() is clockwise


PASS rotate() is clockwise 
